Play touch tone on click and mouseup/mousedown
This allows touch tone to play with pressing the access key, or in the
future when we use better key bindings and manually initiate a click.
There seems to be an issue with the sound playing twice, because the
click event is firing after the mouseUp toggles the flag. I think this
can be fix by a small delay on the toggle, but I'm not sure yet.